fix automatic timezone failing on some GSM networks
in the code that parses the NITZ string, someone made the assumption that if
any data follows the DST flag, it must be be the android emulator appending
an actual timezone string in "Area!Location" format.  However, it seems that
some carriers or perhaps in some locations, the carrier sends an additional
field in the NITZ string.  For example, this is a string from AT&T:

12/07/01,09:37:43-16,01,310

The extra field is "310" above.  This patch will only assume that the field
in question is a textual timezone if and only if it contains an exclamation
point.  Otherwise, the extra field is ignored and the timezone is parsed
as it should be.

Change-Id: I6cc4278027a6de5545d76b803f5359923e60878a
Signed-off-by: Gary Dezern <garyd9@gmail.com>