Locale: Change language not change storage volume descriptions
Actually, if user change language, the volume storage descriptions (used in various places, like
Settings, Camera, ...) are not changed (because StorageVolume references are cached) and still shown
the description of the old language, after phone is rebooted where reference are cached again.

This change adds a method in ServiceManager to notify configuration changes to services that
implements the ConfigurationService interface. Also, it changes MountService to extend of this interface
and refresh the volume cached references with the new translated description label.
In ActivityManagerService, previously to notify the changes to all processes, invokes the new
method of ServiceManager for propagate the notification change.

Change-Id: Ie2604d6b14ddf125fb59e9366c325e4b572df38a